feat: Add dark theme styles for button, filters, and scrollbar
This commit adds styles for the dark theme to the button, filters, and scrollbar. The .dark-theme class is used to target elements specifically when the dark theme is active. The background-color property is used to change the background color of the elements to the --container-color variable value. The ::-webkit-scrollbar and ::-webkit-scrollbar-thumb selectors are used to style the scrollbar for the dark theme, with different shades of the --hue variable used for background color depending on the hover state.